fix a problem with sector refreshing delays at startup
  I misread the original initial values for `RefreshX` and `RefreshY`
from the binary which caused the behavior, and a confused comment from
me. Thanks to Gustinos from OTLand for catching that.
  I actualy done a small back of the envelope calculation and it seems
that, using a config similar to the one from the leaked files, it would
have taken ~100 hours for it to start refreshing sectors, meaning this
feature has NEVER actually been tested. I'd expect unexpected bugs to
emerge from this change.